Bugfix Update #1
Fixed the army_morale_factor modifier in adaptive_learning to actually work and not be misspelled.

Fixed NCR not being able to switch to total mobalized as a focus was named wrong in the trigger.

Fixed Generic Raider focus not giving the power armor equipment from focus.

Fixed Reno having abandon in their focus tree as it did nothing and put stuff that should do the same thing. Also changed some stuff so the if in newcapital focus for enclave actually did something instead of just being there for shits and giggles.

Fixed vegas focus tree complaining about having a invalid idea for papa_khan was just missing a equals.

Not a fix but changed the focus for a valid_war_triggers.txt scripted trigger as the focus in there was commented out so put something else that did the same thing.

Fixed incorrectly spelled war goals in hooverdam.txt and nf_mormons.txt events.

Fixed a misspelled transfer_troops in the april fools events.