Bugfix Update #2 - Little Brother
Made it so only the Western BOS can see their custom conscription laws. Not sure if a bug or something but figured this would be a nice addition.

Made it so a Crucified Lucius can't be in your cabinet or a military leader.

Fixed a description for the vault city focus vault_brain_1 not showing due to it being misspelled.

Localization is for a new custom_trigger_tooltip to inform people of Lucius not being available after crucification.