Implementing replace_path
Implemented the new replace_path feature from 1.5 patch so we can delete all these 0 KB files.

Also deleted all the 0KB files I found.

Reminder to tell the people in the public discord to use the dependency line added in 1.5.2 if they want their submods to work with the mod when this goes live.