Reno Enclave Change
Changed the New Reno Elite party to "The Enclave" after the enclave event happens. Also changed the event to give 100% elite popularity since all the families would have been murdered in that event.

Not sure if we should change it back but rename the other ideologies to something else so they better fit a Enclave Reno.