Fixes/Changes
Preliminary changes to the BOS tree to deal with the Shi expansion. The whole "Trouble on the Homefront" focus branch needs a rework to fit with the new states.

Minor spacing changes to the BOS event file.

Spelling/Syntax corrections to a few states.

Fixed the Buildings.txt and such yelling about building errors for states.