Fixes
Made it so for the Brotherhood NCR war to occur the BOS must exist.

Changed the BOS tree to force the BOS to have the states for focuses that put buildings in them.

Added the requirement of owning the states to the Vault City focuses as well.

Other fixes are spacing changes to be more consistent with the coding style.