Fixes
Made it so the Civil War King Faction from New Vegas gets the flag for the Fourth Family focus tree branch as it was only set in the assassinate house focus for them and no where else so a civil war path could not get the fourth family branch. Change back if this was intentional just didn't seem so to me.

Random space removal for legion event.