More Fixes.
Removed all the errors from ideas/events having picture = without a picture linked by just commenting them out. Needs images in the future.

Changed the duplicate genericbosbunkerthought idea into the non existent genericbossurfaceoutposts idea to remover errors. Will need to be checked to see if it does that genericbossurfaceoutposts is suppose to do.

Added in missing sub-ideology for the Kamlo Survivalists leader Ranger Kenneth.

Changed Aztex event ID's as there was a duplicate nf_aztec.0.

Changed Washington and Mojave Brotherhood first three focuses to be more in line with all the other trees first three focuses. Went from 70 days cost to 7 day cost and changed it so the event is giving instantly instead of having to wait a day.

Fixed errors in syntax for the Alaskan Expedition events.

Removed extra Curly Bracket in the Mojave focus tree.

Removed extra version of Hardin and changed The First's id for troll warren from 982 to 984 to not conflict with Hardin's id.

Changed focus_MOJ_l_english.yml to be in UTF8-BOM encoding.