Legion Focus Change
Change the way the "Calling upon the Nomads" focus is available. Change it so there is no bypass for those two focuses but instead You can only take the "Reward the Loyal Tribe" If the white legs exist and have accepted and you can take the "Calling Upon the Nomads" focus if either the White Legs don't exist OR they exist and refused you.