Fixes
Moved the Mechanized Training infantry training to the advanced side of the infantry tree as it is a advanced tech.

Made it so a NCR trooper doesn't spawn in the ocean.

Gave HEA generic leader/general/admiral portraits.

Did some more if/else if stuff for the NCREclave Civil war stuff. Trying to account for more situations were the Renoclave AI is in a faction or leader of a faction so they properly join the NCRClave faction.