Fixes
Added in localization for the three state countries in the legion that have cores so if people release them they will actually have a name instead of just being blank.

Also gave them the settler focus tree as it seems more approriate for the large state nations then the tribal one.