Changes
Fixed some random spacing/tab issues I found.

Changed the Cult of Nas guys to be motorized instead of APC as having them be APC without having unlocked APC tech for the legion caused a whole bunch of issues.

Fixed the Cult Of Nas Tooltip so it says Motorized instead of APC units.