Changes
Added some events to the NCR-Legion Hoover Dam war so that certain events notified people of losing when they didn't use to notify them.

Moved the Hoover Dam Events localisation from the NCR events to it's own file as it was both the NCR and Legion who did them not just the NCR.

Changed Spacing in a file to be tabs instead of spaces.