GUI for Happy
Main thing is adding a new simple scripted gui for the New Cannon conversion decisions to show if the state is converted or not.

Merged some of the coring_button files like the gfx and localisation to be a scripted_guis file as I didn't see why both needed their own currently as they are both only 2-3 lines in the YML and 1-2 sprites in the gfx.

Added a placeholder gfx for the conversions that we will probably need to replace sooner/later just like the coring button graphic.

Fixed errors from the Mojave news events as they didn't have a namespace or title/desc assigned.

Tiny spacing fixes.