Joshua Graham Icon for NCN
Messed with the Icon indy made added it into the game and changed out the vanilla offensive trait for a custom one that doesn't have a sprite attached.

Left Mag's icon in the gfx folder and left the code to set it up in the .gfx file incase we want to switch back to it.