legionsplosion prelimiary Triggers and more.
Added in the Legionsplosion trigger to get it fired up. This is just the first iteration and may need edits/changes.

Added a news event to make sure we know when the NCR Legion war is starting so the event for NCR can't trigger after every war. Added loc for it in the ncr_event loc file.

Moved around the Legionsplosion scripted effects to it's own file.

Edited the way the nations are released so if they are puppets they won't be puppets anymore.