Vegas/NCR Wargoal changes
Changed the way the NCR declares war on the Mojave Raiders if AI so it doesn't delcare on a non-existant Mojave Raiders.

Changed the way wargoals are created for the New Vegas focuses so it also takes into account if another nation has that territory already.

Added in the AI strategy to keep the Dam Busters in the Vegas faction.

Changed the pictures for some of the generals created by event so they work right.

Spacing fixes.