Spacing fixes, Temp Loc, and Tribe Rebel Stuff
Spacing fixes in a few files I found.

Made the event for the NCR to demand stuff from the Mojave fire_only_once so it stopped spamming me as the NCR.

Changed the "tag" trigger for the news events to be "original_tag" so if the Legion had their civil war it would still trigger.