Spacing and wargoal fixes
Changed Formatting on most of the files to fit the style we have with tabs.

Also added the limit for the wargoals that evil changed to all the trees were applicable.

Moved the custom_effect_tooltips from the BOS tree into a actual localisation file from just being in the focus tree.