Fixed a Error in BRK events. Changed Legion focuses.
Changed a manpower = in the BRK events to be add_manpower as that seemed to be what it meant and manpower isn't a proper effect.

Changed the whole legion Vegas branch to require them to own New Vegas after the original focus. Either by conquering them with the war goal from the original focus or having it bypass and just conquering the land.