Legionsplosion stuff.
Added in a new option for the NCR to switch to the Phoenix Republic when the revolts go off.

Also added in another option in the news events so that the Rebelling nations have their own option. Might need tweaking/additions.

Formating changes in the BRK focus tree. Nothing major.