Event loc placeholder and changes.
Formating/Spacing for some files.

Removed nf_qol.5.b option for nf_qol.5 because there wasn't a reason the BOS wouldn't revolt against you based on the title and description of that event.

Commented out nf_reno.2.a option till factionevent.1 is created and working.

Added placeholder localisation for most of the things that needed it on evil_cookie's list in localisation channel in dev discord.

Placeholder localisation is between the lines placeholder localisation beings and placeholder localisation ends.