First Wave of Neutrality Idea and More.
First wave of adding the neutrality idea. Did it to specific nations to stop early NCR-Legion war and such things.

Formating and Spacing in several files.

Changed Legionsplosion trigger to better account for weird things like going to war with a legion in the civil war.

Added the vanilla placeholder gfx to remove some errors I kept seeing.

Added localisation for the mojave_chapter_chose_leader flag.

Added localisation for the certintyknights idea for BOS.

Fixed error with ACE for New Reno oldbirds OOB.

Added in the missing mormoncrazyhorns OOB from their focus.

Quick fix for a crash from the heaven's gate if Ires or other nation NCN triues to puppet is in a faction with them already. Needs more work to make sure everything is accounted for and works correctly.