Changed "Annex Hidden Valley" legion focus
I changed the Annex Hidden Valley focus for the legion to make more sense with the description were it use to require the Mojave to exist to get the cores I changed it to require them not to exist to take it and give cores on all states they had cored.

As the description talked about them finally being gone and the effect required them to exist in order to get it.